[refactor] NS-39 좋아요 기능 커스텀 훅으로 변경
